cli: extract password io helpers
