+++ crates/cli/src/lib.rs
@@ -129,9 +129,12 @@ use rsync_protocol::{ParseProtocolVersionErrorKind, ProtocolVersion};
 use time::{OffsetDateTime, format_description::FormatItem, macros::format_description};
 use users::{get_group_by_gid, get_group_by_name, get_user_by_name, get_user_by_uid, gid_t, uid_t};
 
+mod password;
 mod progress;
 
 pub(crate) use progress::*;
+
+use password::{load_optional_password, load_password_file};
 
 /// Maximum exit code representable by a Unix process.
 const MAX_EXIT_CODE: i32 = u8::MAX as i32;
@@ -6128,88 +6131,6 @@ fn set_filter_stdin_input(data: Vec<u8>) {
     FILTER_STDIN_INPUT.with(|slot| *slot.borrow_mut() = Some(data));
 }
 
-#[cfg(test)]
-thread_local! {
-    static PASSWORD_STDIN_INPUT: std::cell::RefCell<Option<Vec<u8>>> = const {
-        std::cell::RefCell::new(None)
-    };
-}
-
-#[cfg(test)]
-fn take_password_stdin_input() -> Option<Vec<u8>> {
-    PASSWORD_STDIN_INPUT.with(|slot| slot.borrow_mut().take())
-}
-
-#[cfg(test)]
-fn set_password_stdin_input(data: Vec<u8>) {
-    PASSWORD_STDIN_INPUT.with(|slot| *slot.borrow_mut() = Some(data));
-}
-
-fn load_optional_password(path: Option<&Path>) -> Result<Option<Vec<u8>>, Message> {
-    match path {
-        Some(path) => load_password_file(path).map(Some),
-        None => Ok(None),
-    }
-}
-
-fn load_password_file(path: &Path) -> Result<Vec<u8>, Message> {
-    if path == Path::new("-") {
-        return read_password_from_stdin().map_err(|error| {
-            rsync_error!(
-                1,
-                format!("failed to read password from standard input: {}", error)
-            )
-            .with_role(Role::Client)
-        });
-    }
-
-    let display = path.display();
-    let metadata = fs::metadata(path).map_err(|error| {
-        rsync_error!(
-            1,
-            format!("failed to access password file '{}': {}", display, error)
-        )
-        .with_role(Role::Client)
-    })?;
-
-    if !metadata.is_file() {
-        return Err(rsync_error!(
-            1,
-            format!("password file '{}' must be a regular file", display)
-        )
-        .with_role(Role::Client));
-    }
-
-    #[cfg(unix)]
-    {
-        let mode = metadata.permissions().mode();
-        if mode & 0o077 != 0 {
-            return Err(
-                rsync_error!(
-                    1,
-                    format!(
-                        "password file '{}' must not be accessible to group or others (expected permissions 0600)",
-                        display
-                    )
-                )
-                .with_role(Role::Client),
-            );
-        }
-    }
-
-    let mut bytes = fs::read(path).map_err(|error| {
-        rsync_error!(
-            1,
-            format!("failed to read password file '{}': {}", display, error)
-        )
-        .with_role(Role::Client)
-    })?;
-
-    trim_trailing_newlines(&mut bytes);
-
-    Ok(bytes)
-}
-
 fn parse_bind_address_argument(value: &OsStr) -> Result<BindAddress, Message> {
     if value.is_empty() {
         return Err(rsync_error!(1, "--address requires a non-empty value").with_role(Role::Client));
@@ -6529,30 +6450,6 @@ fn push_file_list_entry(bytes: &[u8], entries: &mut Vec<OsString>) {
     }
 }
 
-fn read_password_from_stdin() -> io::Result<Vec<u8>> {
-    #[cfg(test)]
-    if let Some(bytes) = take_password_stdin_input() {
-        let mut cursor = std::io::Cursor::new(bytes);
-        return read_password_from_reader(&mut cursor);
-    }
-
-    let mut stdin = io::stdin().lock();
-    read_password_from_reader(&mut stdin)
-}
-
-fn read_password_from_reader<R: Read>(reader: &mut R) -> io::Result<Vec<u8>> {
-    let mut bytes = Vec::new();
-    reader.read_to_end(&mut bytes)?;
-    trim_trailing_newlines(&mut bytes);
-    Ok(bytes)
-}
-
-fn trim_trailing_newlines(bytes: &mut Vec<u8>) {
-    while matches!(bytes.last(), Some(b'\n' | b'\r')) {
-        bytes.pop();
-    }
-}
-
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 enum CompressLevelArg {
     Disable,
@@ -6743,6 +6640,7 @@ fn render_module_list<W: Write, E: Write>(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::password::set_password_stdin_input;
     use rsync_checksums::strong::Md5;
     use rsync_core::{branding::manifest, client::FilterRuleKind};
     use rsync_daemon as daemon_cli;
